feat(subscriptions): Allow scheduling watermark mode to be overrideen
This is a temporary change that allows the scheduling watermark mode
to be overridden in the combined scheduler/executor. This is needed for
single tenant as we often have empty partitions there and this is incompatible
with "global" mode. Once transactions are no longer semantically partitioned,
this will not be necessary and this can be removed.