feat: Don't pass next_offset to message (#2792)
We shouldn't ever manually pass a next_offset to Message. This is
because next_offset must always equal to offset + 1, and this will
be properly applied as long as we don't pass anything. The
goal of this change is to remove the ability to pass invalid input into
a message.

In a future version of Arroyo, the ability to pass this at all
will be removed.