fix: Flaky redis test
Digging into this test, everything seemed to be working correctly when the test
failed. My current theory is that there is a small flake in the test that ends
up with the same result.

Either the last slow worker can timeout and find an error in redis, or it can
time out and find nothing in redis. In the first case, it would surface an error
that would be correctly caught by the test. In the second case, it was returning
an error that wasn't correctly handled by the test. Both cases are valid for the
test (slow worker times out waiting for a result from the original query) but
slightly different exceptions.

Since the way `pytest.raises` was not correct, fixing it to correctly watch for
a type seems like a valid fix to me.