fix(replacements) Replace round robin with in order load balancing (#2932)
The current load balancer implementation for replacements uses a round robin algorithm.

This implies that, if we receive two replacements in sequence they are almost certainly going to be executed on different replicas of the same shard.
This works most of the time as having multiple overlapping replacements in a rows is rare, though when this happens we cannot guarantee in order execution of the replacements as the second may end up in zookeeper before the first. If, instead we perform in order load balancing we should always apply the replacement to the same node and we could end up out of order only during failover situations.

This replaces the RoundRobinLoadBalancer with an InOrderLoadBalancer.

This one should go as well as soon as we are on 21.8 everywhere so we can let Clickhouse do the load balancing as, in 21.8, we can issue a query for a specific shard to the distributed node.