feat: Add a str default to JSON encoding
There are some types of data (IP addresses, UUIDs) that are converted to
classes in the result (ipaddress.IPv4Address, uuid.UUID). These classes are not
JSON serializable, so would throw an exception when encoded into the cache or
returned in the JSON response. Instead, automatically encode them as strings.

Context:
In previous datasets (errors, transactions) that had these types of columns
there are mappers and query processors that transform the query so that
Clickhouse returns a string instead of the class. With newer datasets like
replays they do not have these processors. Instead of forcing new datasets to
have these processors, or to wrap all calls with `toString`, we will
automatically JSON encode these results as strings.