fix(optimize): Make optimize jobs stateful (#2571)
Introduce a mechanism of saving state of optimize jobs. This is done by saving the partitions which have already been optimized in redis. The optimized partitions are stored as redis sets. When the job starts it first gets the list of completed partitions from redis. Then it does a diff of partitions which need to be optimized with the partitions which have already been optimized. It then proceeds to perform the optimization on the diff'd set of partitions.

The PR also introduces a notion of threaded cutoff time. When an optimization is running in threaded mode we want to avoid running parallel optimizations when peak traffic starts. The threaded implementation uses the threaded cut off time to run the optimizations. Once all the threads work is done we check again to see if any partitions remain to be optimized using the saved state in redis. The optimization is then performed on remaining partitions using the original cutoff time.