fix(optimize): Make optimize queries non retryable (#2949)
Firing an optimize query causes a merge operation. This operation runs
in the background. So even if an error is raised(which is socket timeout
exception caused because optimize queries take longer to exectue), it
does not indicate a failure in running the query. The cost of re executing
the query is huge. Hence we add an option to specify whether a query should
be retried or not. For optimize queries, we make them non-retryable.

Also increased the timeout value for the optimize queries to 4 hours
since we are seeing the socket timeout exception happen regularly
during beginning of the week.