fix(optimize): Sort partitions each time when getting schedule (#2953)
One of the reasons the optimize jobs are timing out is that there can be overlaps between the TTL merge and optimize merge. When this happens, the other merge is blocked with a reason along the following lines
Not executing log_entry <x> of type MERGE_PARTS because part <y> is not ready yet

TTL merges work with the partition which is 13 weeks old. If we order the partitions to optimize in such a manner that the 13 week old partition is merged in the end, we decrease the likelihood of overlaps happening and optimize jobs timing out