fix(optimize): Remove exception from main execution
The `optimize_partitions` api is mainly called from `optimize_partition_runner`
to control the execution based on different cutoff times and parallelism.
When a timeout is provided to the api, it means that the current schedule can
run only till the provided timeout.

Because `optimize_partitions` is run in a threaded context, it is not easy
to capture its exception since by the time join is called on the thread,
the threads stack trace is already gone.

What is important for us is that the optimize job raise an exception when
OPTIMIZE_JOB_CUTOFF_TIME is reached. Which it does in OptimizeScheduler's
get_next_schedule. We don't care about transitions between different schedules.