fix(replacements) Move tag condition out of the PREWHERE clause (#2942)
In tag deletion we cannot select the rows to replace by running the tag condition in the PREWHERE clause.
The query today is something like

SELECT ....
PREWHERE has(tags.key, 'asd')
FROM errors_dist FINAL
WHERE 
    project_id = .... AND
    ....
This won't work if we replace twice the same row by removing the same tag multiple times before a merge (that would squash the duplicates) happens.
This can happen if you try to delete the same tag twice

The issue is that the FINAL clause of the query would squash the duplicates on the fly and exclude the old row that has already been replaced.
Unfortunately the tag condition is in the PREWHERE that is processed before the FINAL clause. Thus, when running the SELECT, only the old rows makes it to the FINAL clause that has nothing to squash.

This moves the tag condition to the WHERE. It makes the query slower but at least correct.