feat: Produce attribution data to the kafka topic (#2693)
This will produce attribution data to a new Kafka topic, ultimately to be
consumed by a super-big-consumer and published to PubSub. This collects all the
bits of data we care about, puts them into a new container and publishes it to
Datadog and Kafka.

This topic has already been created by the ops team.