feat: Sort raw sql fields and conditions [Experiment]
Context:
To help improve the frequency of raw sql cache hits, this PR is
responsible for sorting column fields and expression conditions in raw
SQL queries. Note: this code should not interferer with the current
operations of Snuba query caching. This is an experiment to investigate
the consistency of the natural ordering of columns and conditions.

Test Plan:
* Provide functionality for sorting, and select a certain percentage
  (defined in runtime config) of raw queries to be part of the
  experiment.
* Exclude queries from an explicit set of referrers to avoid
  some computational overhead.
* Record metrics of cache hits of sorted and unsorted queries.
* Determine whether or not sorting should be implemented for all queries.