feat(optimize): Add jitter to optimize thread start time (#2994)
With parallel optimizations, one observation made is that the load
gets higher during the end of merge. This could potentially be
because theres a lot of IOPS we do during the end of the merge.
With parallel merges, it can get exacerbated because there's a
burst of higher IOPS happening during the end of the merge.

Have added a mechanism where we add a jitter to the starting of
parallel optimize threads to avoid those bursts.