ref(MDC): cleanup dataset factory, formalize factory pattern (#3051)
YAML datasets and python datasets will need to coexist for some time in production. 
The factory functions are the gateway to accessing datasets.

In order to make the transition easier, a few things were refactored in the dataset factory:

1. One container for all global state in the factory (down from four)
2. Only one list of dataset names maintained
3. remove complicated mapping/caching logic


This PR also introduces the ConfigComponentFactory class that I would like to use for all other ConfigComponents (Datasets, Entities, Storages, QueryProcessors) to make the loading and merging of ConfigComponents have a uniform, predictable interface.