ref(MDC): Represent migrations in configuration
This changes the migration groups to load automatically from configuration files
and populate the various global dictionaries we are currently using.

There is a new `ConfigurationLoader` which reads a migration group from a config
path. That path is detected automatically. The loaders are then added to the
global list of loaders and can be accessed in the same method as the hardcoded
loaders.

It is also showing one way to solve the "Enum" problem we have with different
components of Snuba.