refs: Remove `EventsProcessor` and `EventsProcessorBase`
Was reading through this code to understand error processing, and found the two processors a bit
confusing. It seems like `EventsProcessor` is now completely unused, so I removed it.

It seems like `EventsProcessorBase` is also unnecessary now, since we only have `ErrorsProcessor`
subclassing it, so also removed that and consolidataed all into `ErrorsProcessor`.

It could make sense to move the various enums and type defintions out of the `events_processor_base`
file as well. Can do so in this pr or a follow-up if people agree.