fix(grouping): Group clickhouse errors by referrer
Discussed with evan, we want to split up ClickhouseError into smaller
groups that can be ignored individually.

Alternatively we can just start tagging all of those properties, and
use server-side fingerprinting rules (in project settings) to convert
those tags into fingerprints. The advantage of that would be that you
can iterate on grouping much more quickly, but the downside is that
changes to those rules don't live in code and don't undergo a review
process.