ref(MDC): Represent migrations in configuration (#3071)
This changes the migration groups to load automatically from configuration files
and populate the various global dictionaries we are currently using.

There is a new `ConfigurationLoader` which reads a migration group from a config
path. That path is detected automatically. The loaders are then added to the
global list of loaders and can be accessed in the same method as the hardcoded
loaders.

Since the `MigrationGroup` enum was not being used in many places, I replaced
it with simple strings completely, instead of doing the intermediate metaclass.