feat(multistorage-consumer): Add --raw-events-topic and --bootstrap-server options
This brings the `multistorage-consumer` options closer to the `consumer` options.
These two comamnds should eventually be combined anyway, so this is a step in that direction.

The primary motivation for making this change now though is that we are rolling out
errors/transactions split together with the transactions/transactions_v2 double writes.
This requires us to pass non default topic and Kafka cluster options to one of the
multistorage consumers.