fix(errors): Allow input format mismatch
If there is a mismatch between the input format and clickhouse format,
by default clickhouse throws an error. That will lead to ingestion
backlogs in case of format mismatches. Allow the error storage to skip
unknown fields during insertion to make it more robust. This is already
happening when writing to transactions storage.