ref(tests): Use fixed time of 23 hours for optimize test
For most tests around optimize, we never want to run into timeout exception.
That is why we set the timeout to 24 hours. But in the test where we want
an exception, it is probably not a good idea to go into the 24 hour boundary
since it becomes a new key in redis. So use a fixed timedelta of 23 hours
for the test in which we want to test timeouts