fix(TDC) Move TypeConverters into same folder as QueryProcessor (#3253)
The TypeConverter classes were in a separate folder from the rest of the query
processors, since they inherited from BaseTypeConverter, which is a subclass of
QueryProcessor.

This broke the registered class paradigm, since all the children of a registered
class need to be in the same folder. This solution seemed better than creating a
new registered class for type converters, since there is currently no way to
distinguish a type converter from a regular processor in the configuration.