feat: Drop filtering by offset on errors dataset (#3263)
None of the other datasets apply this condition (they cannot since they are non semantically partitioned).

We sometimes flip large projects into to random partitioning mode to avoid backlogs and incidents. Doing so causes the alert to be completely wrong unless we remove this condition as the offset across different partitions can no longer be compared.