fix(metrics): Remove broken statsd metric for rate limiter
`stale_queries` is not a number, because the pipeline didn't execute
yet. In datadog the metric does not exist. It would be easy to fix this
metric but I doubt its usefulness. We already have alerts on growing
redis memory usage that tell us when the rate limiter is misbehaving.