feat(observability): Add measurements to readthrough cache (#3270)
We are having issues with occasionally spiking query times in snuba. We have evidence to believe that this is because of the readthrough cache but we don't know why exactly because its inner workings are not visible at runtime. This PR introduces metrics about the cache's operation such that they can be observed when another spike happens