feat(rate_limit): Shard sets of table rate limiter
Redis sets are not sharded across multiple nodes. The table rate limiter
manages a set that grows proportional to the amount of concurrent
queries allowed + whatever is configured for history purposes.

It's therefore possible for a clickhouse cluster of sufficient size to
allow for much more per-second and concurrent queries than what a single
redis node can handle.

In order to bypass this problem, we shard the redis set by query ID.

We did something similar for the cardinality limiter in Sentry
("sentry.ratelimits.cardinality"), where there's the concept of
"num_shards" and "num_physical_shards".