feat(redis): Add second shard to rate limiter and cache
We want to split out those usecases to a new redis cluster, but instead
of doing a hard cutover to a new redis instance we want to do it slowly.

We don't care much about dataloss (particularly for those usecases), our
primary concern is whether the new Redis instance can take the load of
those usecases. We have little insight into what causes the most load on
the current redis cluster, and at the same time hope to get away with a
single memorystore instance per-usecase as memorystore does not support
sharding yet.

In order to facilitate a "slow" cutover we introduce two new rollout
rate settings. Increasing a rollout rate will make the corresponding
usecase gradually start a different redis cluster (represented as a new
"v2 usecase"), and once we hit 1.0 on both of them we can essentially
revert this PR and point the "v1" usecase directly to the new cluster
(in one deploy)

There is a chance this doesn't go well and a single memorystore instance
cannot take the load of either the rate limit or caching usecase, in
which case we have to consider building sharding into Snuba.

Longer-term I'd ideally want us to move to Envoy's Redis proxy or a
similar abstraction layer between the redis client and the actual
instances. That would give us dual-writes and key-based sharding OOTB
without having to build it into the app itself.