ci(clickhouse): Enable running tests against clickhouse version 21.8 (#3292)
We have multiple versions of clickhouse running across different environments. Some clusters on SaaS run on version 21.8 while self hosted and single tenant run anywhere from 20.3 to 20.7. The goal of this PR is to ensure that the code we introduce passes on the minimum and maximum versions of clickhouse we support i.e. 20.3 and 21.8. This is done by parametrizing the clickhouse version in the docker compose used by gcb.