feat(optimize): track merge state (#3158)
Addresses [SNS-1483](https://getsentry.atlassian.net/browse/SNS-1483). 

Adds functionality to query `system.merges` at the start of an optimization and delay running `OPTIMIZE` on a partition if there are already ongoing merges on that partition.

It works by calling `optimize.is_busy_merging()` to check if it should execute a merge or sleep then retry later. 

`is_busy_merging` queries `system.merges` table for merges executing and returns true and the estimated sleep time if clickhouse is busy with merges in progress for the table. Clickhouse is considered busy if:
1. there are more than `settings.OPTIMIZE_MERGE_MAX_CONCURRENT_JOBS` merges in progress with an elapsed time > `settings.OPTIMIZE_MERGE_MIN_ELAPSED_CUTTOFF_TIME`. These are merges that we consider long running and should probably wait on before attempting again.
2. or there is a merge where `total_size_bytes_compressed` is greater than `settings.OPTIMIZE_MERGE_SIZE_CUTOFF`. These are merges on large parts and we should wait for them.

In both cases, the sleep time is `setting.OPTIMIZE_BASE_SLEEP_TIME + estimated_sleep_time` where `estimated_sleep_time` is the max of `elapsed/progress` of the ongoing merges found in the query.