Revert "ref(MDC): Represent migrations in configuration (#3071)"
It's not clear that migrations belong in configuration. This was already an
issue from the previous work, since the migrations themselves are still part of
the code. There is now a separate effort to figure out where migrations should
live, so reverting these changes in the mean time to avoid having things in a
half finished state.

Before state:

The generic metrics MigrationGroup was defined in configuration, and the groups
code was loading that configuration. The MigrationGroup was a string to allow
for groups being defined outside of Python code.

After state:

All MigrationGroups are defined in the python file, and MigrationGroup has been
changed back to an Enum.

This reverts commit ae89a81b219889b3dcbe53c18f7f27c35aa4e275