feat: Make multistorage consumer dlq respect the policy defined (#3343)
Currently the multistorage consumer completely ignores the policy defined on the storage itself and just applies it's own policy which produces the message to a topic. This is wrong and also different to how the main consumer works where the policy defined on the storage is the one that is used. It was also applied on a different part of the pipeline (only wraps the collector step). This also means the collector implementation was inconsistent in the two consumer variants.

When this hack was first introduced, it's likely there were practical reasons we needed this at Sentry, but since this particular code path (multistorage consumer with dlq) is never used today, those obviously don't exist anymore.

This is a step on the path to folding the multistorage consumer into the main consumer and making them behave in the same way.