fix(query cache): Retry query with randomized ID if cache fails
Figured I'd open this without tests just to get some feedback on the
approach.

It seems that ClickHouse rejects the second of two queries that have the
same query ID. The query ID is directly computed from the formatted
query contents itself, so two identical queries hitting clickhouse
yields errors.

This is typically not a problem as we use a Redis cache to deduplicate
queries. However, that Redis cache can fail and lose data, such as
during a failover scenario or when I am specifically trying to migrate
the cache from one Redis cluster to another without doing data
migration.

I think something like this is the sanest attempt to avoid rejecting
valid queries without significantly affecting what query IDs mean (in
querylog for example) most of the time. A significantly simpler approach
would be to randomize the query ID that is being passed to clickhouse
while still using the formatted query for the cache key. I don't see a
strong reason for not doing that right now but it might affect auditing
when trying to correlate cache keys in redis with the query IDs in
querylog.