fix(clickhouse): Annotate INSERT error with the relevant row [SNS-1850]
Set the row's payload as Sentry context, so that when the snuba consumer
crashes on INSERT, the relevant sentry error event contains the faulty
payload. Sentry contexts have the most generous size limits and are
least likely to be truncated by Relay. At the same time Relay should be
smart enough to discard this data before it attempts to trim down more
important info, like the error message or stacktrace. We'll see.

I didn't find a good place to grab this data. It seems that we stream
the rows from processing strategies without ever holding all of them in
memory. As such it's very likely that this patch significantly increases
memory consumption, because we need to buffer up all encoded rows in
case we need to add them to an error.

Also the rows are encoded bytestrings, so they will look a bit ugly in
Sentry but that's also the cheapest representation to keep in memory
until the end of the batch (one PyObject per message)