ref(querylog) Add pre-defined queries to Querylog Admin
Similar to System Queries, add a selection of pre-defined queries in the
querylog tab with descriptions of what they mean and how to use them.

Add an endpoint that powers the selection, and tests for that endpoint as well
as the system queries equivalent. In order to avoid code duplication, also
change the system queries registry to use the RegisteredClass concept for
registering all the predefined queries. This could then be re-used for the
querylog queries.