fix(migrations) Allow ttl_day on ModifyColumn operations
There is a use case where someone wants to be able to give a daily TTL to a
specific column. The operation currently only allows monthly TTL.

Before:

Columns could only have monthly TTL if they needed a new  TTL from the table.

After:

The ModifyColumn operation takes an optional `ttl_day` parameter that works the
same way as `ttl_month` but on a daily interval instead of monthly.