feat(migrations): Allow TTL clause on columns in migrations
Currently the only way to specify a TTL for a specific column in a table is to
have a ModifyColumn migration to add the TTL after the table is created. It
should also be possible to specify those TTLs in the create table however.

Before:

There was no TTL modifier for a Column, so a ModifyColumn migration had to run
in order to add a TTL to a specific column.

After:

There is now a TTL modifier that can specify a custom TTL expression on a column