make replacer state a string in storage definition (#3479)
## Context

Replacer state is an argument in the errors configuration but we don't want it to be an enum since that's not configurable in yaml. Convert it to an enum inside the init function of the replacer and the PostReplacementConsistencyEnforcer

## Blast Radius

Only the errors dataset is affected, there are no other replacement states defined in the codebase.