ref(abuse) Add more predefined querylog queries (#3489)
Add two new queries to help narrow down with referrers might be causing the
abuse. To use these queries you should already know which cluster is
experiencing the abuse, and at what time the abuse started.

The idea behind the queries is: take half an hour before the abuse started, and
half an hour after. Compare the referrers between those two time periods, and
surface the ones that had the highest percentage change from before to after.

The purpose of these is not to automatically find the smoking gun, but to at
least remove some of the noise and show some actionable signal that can be
followed up on to confirm if one of the referrers actually is the source of
the problem.

I tested these queries on some known cases of abuse, and they did surface the
referrers that turned out to be the problem.