ref: Rewrite the KafkaConsumerWithCommitLog as a strategy
The ``KafkaConsumerWithCommitLog`` was a huge hack that essentially
moved what should be part of the processing strategy logic (produce and
commit) into the consumer code instead.

The main issue with this is that in addition to be a bit against the
library's design, it required a change to Arroyo's interface to even
support - instead of committing simply an offset
and it's partition, we need to pass the original message timestamp
all the way through the commit interface unnecessarily. This makes
every other consumer (where committing never needs the timestamp)
more complicated.

This is the first step towards deprecating the commit log consumer,
and simplifying committing offsets in Arroyo. It removes the
KafkaConsumerWithCommitLog from the main Snuba consumer. It will be
removed from the multistorage consumer as a follow up.