ref: Rewrite the KafkaConsumerWithCommitLog as a strategy (#3506)
The ``KafkaConsumerWithCommitLog`` was a huge hack that essentially moved what should be part of the processing strategy logic (produce and commit) into the consumer code instead to be executed post strategy.

The main issue with this (in addition to be a bit against the library's design) is that it required a big change to Arroyo's interface to support - instead of committing simply an offset and it's partition, we need to now pass the original message timestamp all the way through the commit interface unnecessarily.... just because of this consumer. This makes every other consumer (where committing never needs the timestamp) more complicated.

This is the first step towards deprecating the commit log consumer, and simplifying committing offsets in Arroyo. It removes the KafkaConsumerWithCommitLog from the main Snuba consumer. It will be removed from the multistorage consumer as a follow up.