feat: Deprecate the KafkaConsumerWithCommitLog (#3523)
Another take at killing the KafkaConsumerWithCommitLog. This implementation moves the part that produces offsets to the commit log topic into the collect step.

For context, the previous take on it was https://github.com/getsentry/snuba/pull/3511/ but this other approach doesn't work well. Since Collect/ParallelCollect does not pass any output to the next step, it must be done within this processing step itself.

This change is needed as Arroyo will no longer pass message timestamps through to the commit callback function via `Position`. It is wasteful, affects performance and only existed to support this particular use case (writing part of the strategy logic in the consumer) which is not really a valid one.