feat: Skip invalid offset ranges (#3530)
We had an incident at Sentry (inc-276) caused by the fact that there was an offset/timestamps in the incorrect order in one of the consumers. We are supposed to be catching and ignoring these in the subscriptions scheduler.

We were already handling invalid time intervals, but we forgot to include handling for invalid offset intervals.