ref(abuse) Add the parent_api back to the metrics (#3497)
The parent_api metric can be useful to see which feature has spiked in abuse,
since one feature can call many referrers and make it harder to see changes in
traffic usage.

This metric was removed in the past because we found cases of the endpoint not
being correct. Referrers were attached to endpoints that didn't actually call
the referrer. However the Sentry SDK has improved quite a bit since then and
even if the endpoint is not correct, it can still be useful signal of a problem.