feat: Ensure producer is flushed
We need to ensure the producer is flushed during the consumer shutdown
process.This fix flushes the commit log producer on each batch, which
adds a blocking operation on each batch. Ideally it would only be performed
when the consumer is actually shutting down but the current design of the
consumer (and use of the CollectStep) does not allow us to easily hook into that.