fix: Ensure commit log producer is flushed (#3535)
We need to ensure the producer is flushed during the consumer shutdown process.This fix flushes the commit log producer on each batch. It does add a blocking operation on each batch. Ideally it would only be performed when the consumer is actually shutting down but the current design of the consumer (and use of the CollectStep) does not allow us to easily hook into that. Given the size of batches is fairly large, i think this flush is non problematic.

This is a fix for SNUBA-30Y firing during deployment.