ref(multistorage-consumer): Do not use ConsumerStrategyFactory in multistorage consumer
The KafkaConusmerStrategyFactory is supposed to get deprecated. This stops using it
in the multistorage consumer and removes the need for the generic `ConsumerStrategyFactory`
which takes a MultistorageKafkaPayload instead of a KafkaPayload.

Soon we will be adding a new processing step for payload decoding and schema validation - this
will happen between the DLQ and the rest of the message processing steps.

This change will make it easier to introduce the decoder/validator as it needs to happen
in different places in the main consumer vs the multistorage consumer.