ref: Remove dead code around multistorage consumer identical storages (#3552)
This function doesn't do anything since the IDENTICAL_STORAGES set is empty (and it should always be empty).

This code existed as part of the ClickHouse upgrade strategy at Sentry which involved us switching to multistorage consumers and writing two clusters instead of performing the upgrade directly. Since we are not doing upgrades this way going forward, it is unlikely this will be needed again.