feat(consumers): Remove the co-operative rebalancing option
Co-operative rebalancing never quite worked with the way we write our consumers
and is not very compatible with Arroyo. Let's remove the option since we never want to run this.

Eventually we might want to try this again but will probably have to wait for https://cwiki.apache.org/confluence/display/KAFKA/KIP-848%3A+The+Next+Generation+of+the+Consumer+Rebalance+Protocol
being implemented in Kafka which will be some time away.