feat: Consumer schema validation, take 2 (#3566)
Same as https://github.com/getsentry/snuba/pull/3553 but instead of passing JsonCodec to the `process_message` function we are now passing the topic instead.

The previous implementation caused consumers to crash if multiprocessing was used (INC-279 / SNUBA-316) as JsonCodec cannot be pickled in order to pass to subprocesses. The new implementation loads the JsonCodec once in each subprocess and only passes the topic to the `process_message` function so the correct codec and schema can be loaded.