feat: Deprecate settings.TOPIC_PARTITION_COUNTS
The goal of this change is to reduce the amount of unnecessary configuration in Snuba.
Requiring operators of Snuba to set the right TOPIC_PARTITION_COUNTS values is error prone and
causes additional work and complexity during incidents. Now the KafkaTopicSpec.partitions_number
fetches the real partitions number from the broker when starting up instead of the hardcoded setting.

The KafkaTopicSpec class now only returns actual partition counts instead of counts from configuration.
The implication of this is that it can no longer be used to bootstrap topics that don't exist yet (e.g.
in dev). Topics bootstrapped via the `snuba bootstrap` command are now always created with 1 partition.