feat(oncall): add the request_id to the query breadcrumbs (#3561)
Searching for queries by id in the querylog requires the request_id because that is the orderBy key for the table. Without it, the queries timeout. This assures that any query that makes it to sentry has a request_id attached to it
add the request_id to the query breadcrumbs