fix(DC): Bring EntityContainsColumnsValidator back to config entities (#3540)
Due to a bug in our testing, we have not been including the EntityContainsColumnsValidator in the entities built with config. Having a place where the built in validators are defined makes it easier to auto-convert entities as we can skip the default ones in the config. 

### Blast Radius

This will affect only config entities (metrics) and this is re-introducing a columns validator which is turned off. This is essentially a no-op.