ref(DC): Reformat how entities specify storages in configuration (#3591)
### Summary
Currently, we are not able to move the `sessions` entity into configuration. This is because sessions contains different translation mappers for each of its selectable storages. This PR is responsible for reformatting how storages and translation mappers are defined in configuration. Entities now specify a `Sequence[EntityStorageConnection]`. In addition, a `SimpleQueryStorageSelector` was added to support this new format of storages in configuration.

### Before State
Entity configuration:
```
readable_storage: generic_metrics_distributions
writable_storage: generic_metrics_distributions_raw
translation_mappers: [x,y,z]
```
* Storages were "implicitly" selected in code. For example, when there are two or more storages defined in an entity, only one single storage (usually the materialized_storage) is passed into the query plan builder.

### After State
* Entity configuration:
```
storages:
  - storage: generic_metrics_distributions
  translation_mappers: [x,y,z]
  - storage: generic_metrics_distributions_raw
  is_writable: true
  translation_mappers: [x,y,z]
```
* All storages defined in configuration are passed into the query plan builder. It is the job of the query storage selector to determine which storage to use. As a result, all storages need to be selected explicitly in configuration. The addition of  `SimpleQueryStorageSelector` helps with that. 

### Blast Radius
The configuration of all entities have been changed to use this format.

### Testing
* Entity definitions and pluggable entity tests have been updated to use new format
* A new `SimpleQueryStorageSelector` added.