feat(config): Migrate GroupedMessage and GroupAssignee Entities to YAML
GroupedMessage and GroupAssignee entities had YAML versions created and added to
their existing datasets. The two entities are now being loaded from config
instead of Python.

`required_time_column` can now be set to `null`. The field is still required in
the schema since that is the main case. These entities are a special case.

There is now a test in the entity loader that is similar to the storages test
that compares entities to their Python version. As we migrate entities we should
update that test. Once our migration is done we can delete that test entirely.