feat: Remove ability to override scheduling mode (#3623)
Remove the ability to override the scheduling mode on the combined scheduler/executor.
This hack previously existed for single tenant's `transactions` topic. Since the
topic was semantically partitioned but the mode was global, empty partitions would
cause the entire pipeline to stop on a regular basis.

This combination (semantic partitioning with subscriptions in global mode) no longer
exists so this is not needed.