Surface Kafka connection errors (#3598)
INC-287 showed that we do not surface in any way Kafka connection errors except for the librdkafka logs.
In case of persistent errors like when the pod cannot establish a connection for a network problem, on k8s, we should only retry a given amount of times in the same pod, then we should try to restart the pod somewhere else.

This is a first step as it adds an error callback so we can see how often this happens. A followup is to track the amount of failure per time period so we can take actions.