fix(health) Fix extra data in storage set error (#3660)
When an `UndefinedClickhouseCluster` exception was raised sometimes it was
adding extra data to the error. That would be fine except that extra data is
being passed into metrics tags which expect a certain format. Change the error
to correctly pass extra data so it gets propagated to the metric tags correctly.