ref(config) Remove Discover Python components
This also required restructuring some values that existed in the Python
component files. The definition of event / transaction columns for the purposes
of selecting an entity moved to the special discover file, and some tests were
updated to have the values stored directly.

Now that the Discover selection is all in one place, it should be fairly
straightforward to port that function directly to Sentry.